Añadido Nube de Tags. Faltaría probarla exhaustivamente pero parece que va bien. Arreglados diversos errores.
--*Tag*.cs
Cambiado tipo del percent en el tagDto.
Agregado metodo FindById y añadido Load al metodo que lo necesitaba.

--Styles.css
Añadido CSS de la nube de tags.

--PracticaMad.Master*
Añadida nube de tags, tanto el HTML como el código c#

--MainPage
Corregido un pequeño error que hacía que en ciertas ocasiones no apareciera correctamente, un label.

--ViewComments
Añadida la posibilidad de que ademas de los comentarios de un evento, muestre los comentarios etiquetados con un tag.

--ViewComment*.resx
Añadidas referencias sobre una nueva etiqueta.

+++ PracticaMaD/trunk/Model/TagService/TagDto.cs
@@ -25,7 +25,7 @@ namespace Es.Udc.DotNet.PracticaMaD.Model.TagService
         /// <value>
         /// The percent.
         /// </value>
-        public float percent { get; private set; }
+        public double percent { get; private set; }
 
         #endregion
 
@@ -34,7 +34,7 @@ namespace Es.Udc.DotNet.PracticaMaD.Model.TagService
         /// </summary>
         /// <param name="tag">The tag.</param>
         /// <param name="percent">The percent.</param>
-        public TagDto(Tag tag, float percent)
+        public TagDto(Tag tag, double percent)
         {
             this.tag = tag;
             this.percent = percent;
